requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-decouple==3.8
pillow==10.1.0
django-filter==23.3
django-extensions==3.2.3
openai==1.3.5
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.17.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
whitenoise==6.6.0
gunicorn==21.2.0
